Week 2 Lessons 3.2: Get Single Album by ID
We achieved that by using  req.params.albumId and storing the value in a variable called albumIdToFind. Then using the .find method on the albumsData  to find the album that has the matching id or albumIdToFind and then display that single album.

Tested by sending:
localhost:35981/albums/11 to get the album that has 11 as its id.